requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
django-filter>=23.2,<24.0
channels>=4.0.0,<5.0.0
channels-redis>=4.1.0,<5.0.0
daphne>=4.0.0,<5.0.0
Pillow>=10.0.0,<11.0.0
psycopg2-binary>=2.9.6,<3.0.0
python-dotenv>=1.0.0,<2.0.0
drf-yasg>=1.21.7,<2.0.0
django-cors-headers>=4.2.0,<5.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
celery>=5.3.1,<6.0.0
redis>=4.6.0,<5.0.0
django-jazzmin>=2.6.0,<3.0.0 